chore(table) convert to TS
chore(DashboardWrapper) convert to TS

chore(Table) convert demos to TS

delete superfluous dashboard components

declare as ts files

remove duplicate ids

fix import errors, WIP build errors

sync toolbar demo with main
