+++ packages/react-table/src/demos/examples/Compact.tsx
@@ -1,9 +1,7 @@
-import React from 'react';
-
+import React, { useState } from 'react';
 import {
   Button,
   Card,
-  Label,
   MenuToggle,
   MenuToggleElement,
   Pagination,
@@ -13,29 +11,33 @@ import {
   Toolbar,
   ToolbarContent,
   ToolbarGroup,
-  ToolbarItem
+  ToolbarItem,
+  Label,
+  PaginationVariant
 } from '@patternfly/react-core';
 import { Table, TableText, Thead, Tr, Th, Tbody, Td } from '@patternfly/react-table';
 import FilterIcon from '@patternfly/react-icons/dist/esm/icons/filter-icon';
+import { rows, columns } from '@patternfly/react-table/src/demos/examples/Table/sampleData';
 import { DashboardWrapper } from '@patternfly/react-core/src/demos/DashboardWrapper';
-import { rows, columns } from '@patternfly/react-table/src/docs/demos/table-demos/sampleData';
 
-export const CompactTable = () => {
-  const [isSelectOpen, setIsSelectOpen] = React.useState(false);
-  const [page, setPage] = React.useState(1);
-  const [perPage, setPerPage] = React.useState(10);
-  const [paginatedRows, setPaginatedRows] = React.useState(rows.slice(0, 10));
-  const handleSetPage = (_evt, newPage, perPage, startIdx, endIdx) => {
+export const CompactTable: React.FC = () => {
+  const [isSelectOpen, setIsSelectOpen] = useState<boolean>(false);
+  const [page, setPage] = useState<number>(1);
+  const [perPage, setPerPage] = useState<number>(10);
+  const [paginatedRows, setPaginatedRows] = useState(rows.slice(0, 10));
+
+  const handleSetPage = (_evt: any, newPage: number, startIdx: number, endIdx: number) => {
     setPaginatedRows(rows.slice(startIdx, endIdx));
     setPage(newPage);
   };
-  handlePerPageSelect = (_evt, newPerPage, newPage, startIdx, endIdx) => {
+
+  const handlePerPageSelect = (_evt: any, newPerPage: number, newPage: number, startIdx: number, endIdx: number) => {
     setPaginatedRows(rows.slice(startIdx, endIdx));
     setPage(newPage);
     setPerPage(newPerPage);
   };
 
-  const renderPagination = (variant, isCompact) => (
+  const renderPagination = (variant: string, isCompact: boolean) => (
     <Pagination
       isCompact={isCompact}
       itemCount={rows.length}
@@ -43,7 +45,7 @@ export const CompactTable = () => {
       perPage={perPage}
       onSetPage={handleSetPage}
       onPerPageSelect={handlePerPageSelect}
-      variant={variant}
+      variant={variant as PaginationVariant}
       titles={{
         paginationAriaLabel: `${variant} pagination`
       }}
@@ -63,7 +65,7 @@ export const CompactTable = () => {
               </MenuToggle>
             )}
             isOpen={isSelectOpen}
-            onOpenChange={(isOpen) => setIsSelectOpen(isOpen)}
+            onOpenChange={(isOpen: boolean) => setIsSelectOpen(isOpen)}
             onSelect={() => setIsSelectOpen(!isSelectOpen)}
           >
             {[
@@ -92,7 +94,7 @@ export const CompactTable = () => {
     </Toolbar>
   );
 
-  const renderLabel = (labelText) => {
+  const renderLabel = (labelText: string) => {
     switch (labelText) {
       case 'Running':
         return <Label color="green">{labelText}</Label>;
@@ -102,8 +104,11 @@ export const CompactTable = () => {
         return <Label color="blue">{labelText}</Label>;
       case 'Down':
         return <Label color="red">{labelText}</Label>;
+      default:
+        return <></>;
     }
   };
+
   return (
     <React.Fragment>
       <DashboardWrapper hasPageTemplateTitle>
@@ -146,7 +151,6 @@ export const CompactTable = () => {
                 ))}
               </Tbody>
             </Table>
-            {renderPagination('bottom', false)}
           </Card>
         </PageSection>
       </DashboardWrapper>
