+++ packages/react-table/src/demos/index.ts
@@ -1,0 +1,1 @@
+export * from './sampleData';
